Move tests to temp
