+++ __tests__/jest_setup/setup.js
@@ -3,7 +3,7 @@ const { client } = require("../../src/index.js");
 const BOT_TOKEN = process.env.BOT_TOKEN;
 
 beforeAll(async () => {
-  await client.login(BOT_TOKEN);
+  // await client.login(BOT_TOKEN);
 });
 
 // Find all test files in __tests__ root
